+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: cdana <cdana@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/11/30 16:46:30 by cdana             #+#    #+#             */
-/*   Updated: 2020/01/03 12:29:23 by cdana            ###   ########.fr       */
+/*   Updated: 2020/01/07 12:33:20 by cdana            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,6 +15,7 @@
 # include <stdarg.h>
 # include <stdlib.h>
 # include <unistd.h>
+# include <stdio.h>
 
 typedef struct	s_arg
 {
@@ -35,7 +36,7 @@ int				ft_print_x(t_arg *arg, va_list ap);
 int				ft_print_xx(t_arg *arg, va_list ap);
 int				ft_printf(const char *s, ...);
 int				ft_atoi(const char *s);
-int				ft_itoa_base(char *base, long long nb, char **line);
+int				ft_itoa_base(char *base, unsigned long nb, char **line);
 int				ft_printarg(const char *s, int *i, va_list ap);
 int				ft_fill_struct(const char *s, t_arg *arg, va_list ap);
 #endif
